fully working, clean solution for Live streaming

+++ ProcessStreamer/StreamConfig.cs
@@ -5,6 +5,6 @@ namespace ProcessStreamer
     {
 		public string Link { get; set; }
 		public string Name { get; set; }
-		public int ChunkTime { get; set; }
+		public double ChunkTime { get; set; }
     }
 }
